backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
qrcode==7.4.2
pillow==10.0.1
paramiko==3.3.1
cryptography==41.0.7 